Fix test-fixture module imports
