+++ backend/search/src/jsMain/kotlin/com/gchristov/thecodinglove/search/usecase/RealSearchWithHistoryUseCase.kt
@@ -3,31 +3,14 @@ package com.gchristov.thecodinglove.search.usecase
 import arrow.core.Either
 import arrow.core.flatMap
 import com.gchristov.thecodinglove.searchdata.SearchRepository
-import com.gchristov.thecodinglove.searchdata.model.Post
 import com.gchristov.thecodinglove.searchdata.model.SearchConfig
 import com.gchristov.thecodinglove.searchdata.model.SearchError
+import com.gchristov.thecodinglove.searchdata.usecase.*
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import kotlin.random.Random
 
-interface SearchWithHistoryUseCase {
-    suspend operator fun invoke(
-        query: String,
-        totalPosts: Int? = null,
-        searchHistory: Map<Int, List<Int>>,
-    ) : Either<SearchError, Result>
-
-    data class Result(
-        val query: String,
-        val totalPosts: Int,
-        val post: Post,
-        val postPage: Int,
-        val postIndexOnPage: Int,
-        val postPageSize: Int
-    )
-}
-
-internal class RealSearchWithHistoryUseCase(
+class RealSearchWithHistoryUseCase(
     private val dispatcher: CoroutineDispatcher,
     private val searchRepository: SearchRepository,
     private val searchConfig: SearchConfig
