+++ backend/slack-data/src/jsMain/kotlin/com/gchristov/thecodinglove/slackdata/usecase/VerifySlackRequestUseCase.kt
@@ -1,0 +1,16 @@
+package com.gchristov.thecodinglove.slackdata.usecase
+
+import arrow.core.Either
+import com.gchristov.thecodinglove.commonservicedata.api.ApiRequest
+
+interface VerifySlackRequestUseCase {
+    suspend operator fun invoke(request: ApiRequest): Either<Throwable, Unit>
+
+    sealed class Error(message: String? = null) : Throwable(message) {
+        object MissingTimestamp : Error()
+        object MissingSignature : Error()
+        object TooOld : Error()
+        object SignatureMismatch : Error()
+        data class Other(override val message: String?) : Error(message)
+    }
+}
